requirements.txt
torch>=2.0
torch-geometric>=2.4
numpy>=1.24
pandas>=2.0
scikit-learn>=1.3
matplotlib>=3.7
seaborn>=0.12
pyyaml>=6.0
networkx>=3.0
umap-learn>=0.5
